Feat: Implements Posts in home page
